source/decode_publish_dbg/requirements.txt
numpy
pandas
mysql-connector
paho-mqtt
